Fix: overwrite prompt was hidden
The logging strategy for FFmpeg (hide all outputs and write all to a log file, and let the user see the log file on error) had the unfortunate consequence of hiding the overwrite prompt when an output file already existed. That strategy was abandoned. It would be great to have two FFmpeg logging levels: one for the display and one for the report.